Update to most recent version of stj-schema-mapper
